update mock paint

+++ te/src/test/java/me/chan/te/test/mock/MockTextPaint.java
@@ -17,7 +17,7 @@ public class MockTextPaint extends TextPaint {
 	}
 
 	public int getMockTextHeight() {
-		return (int) (mMockTextSize / 2);
+		return mMockTextSize / 2;
 	}
 
 	@Override
